Made sure that n_rz_azimuthal_modes is declared int everywhere

+++ Source/Particles/PhysicalParticleContainer.cpp
@@ -578,7 +578,7 @@ PhysicalParticleContainer::AddPlasma (int lev, RealBox part_realbox)
     Real density_max = plasma_injector->density_max;
 
 #ifdef WARPX_DIM_RZ
-    const long nmodes = WarpX::n_rz_azimuthal_modes;
+    const int nmodes = WarpX::n_rz_azimuthal_modes;
     bool radially_weighted = plasma_injector->radially_weighted;
 #endif
